+++ src/treecipe/src/RecipeService/RecipeService.ts
@@ -1,10 +1,10 @@
-import { IFakerService } from "../FakerService/IFakerService";
+import { IRecipeFakerService } from "../RecipeFakerService.ts/IRecipeFakerService";
 import { RecordTypeWrapper } from "../RecordTypeService/RecordTypesWrapper";
 import { XMLFieldDetail } from "../XMLProcessingService/XMLFieldDetail";
 
 export class RecipeService {
 
-    private fakerService: IFakerService;
+    private fakerService: IRecipeFakerService;
     constructor(private fakerServiceImplementation) {
         this.fakerService = fakerServiceImplementation;
     }
